feat(register_notification) configured rabbitmq in auth microservice
